refactor: Relocate and restructure util classes
This commit refactors the project structure by relocating utility classes, specifically mappers and enums, into more appropriate layers to better align with clean architecture principles. The `utils` package has been removed, and its contents are now organized under the `domain` and `ui` layers.

**Key Changes:**

*   **Package Restructuring:**
    *   The top-level `utils` package has been deleted.
    *   Its contents have been moved into `domain/utils` and `ui/utils` packages.

*   **Mapper Relocation:**
    *   `Mapper.kt` (from `utils/extension`) has been split and relocated:
        *   `ProductMapper.kt` is a new file in `domain/utils/mapper`, containing the domain-level `toMerge` extension function.
        *   `UiMapper.kt` is a new file in `ui/utils/mapper`, containing UI-related mapping functions like `toFavoriteLocalModel`, `toProductLocalModel`, and `toOrdersUiModel`.
    *   `toProductEntity()` has been renamed to `toProductLocalModel()` to more accurately reflect the model it maps to.

*   **Enum and Util Relocation:**
    *   `OrdersProcessEnum.kt` moved from `utils/enums` to `domain/utils/enums`.
    *   `Util.kt` (containing `dpToPx`) moved from `utils/util` to `ui/utils`.

*   **Import Updates:**
    *   All necessary import statements across ViewModels (`SharedViewModel`, `BasketViewModel`, `DetailViewModel`, `HomeFragmentViewModel`), UseCases (`GetProductUseCase`, `UpdateOrdersUseCase`), and UI components (`Basket.kt`) have been updated to reflect the new locations of the mapper functions and enums.
