+++ app/src/main/java/com/example/e_commerce/domain/usecase/GetProductUseCase.kt
@@ -3,7 +3,7 @@ package com.example.e_commerce.domain.usecase
 import com.example.e_commerce.domain.model.ProductsModel
 import com.example.e_commerce.domain.repository.local.LocalProductRepository
 import com.example.e_commerce.domain.repository.remote.ProductRepository
-import com.example.e_commerce.utils.extension.toMerge
+import com.example.e_commerce.domain.utils.mapper.toMerge
 import javax.inject.Inject
 
 class GetProductUseCase @Inject constructor(
