+++ app/src/main/java/com/example/e_commerce/domain/utils/enums/OrdersProcessEnum.kt
@@ -1,6 +1,7 @@
-package com.example.e_commerce.utils.enums
+package com.example.e_commerce.domain.utils.enums
 
 enum class OrdersProcessEnum {
     MINUS,
     PLUS,
 }
+
