+++ app/src/main/java/com/example/e_commerce/ui/home/HomeFragmentViewModel.kt
@@ -6,7 +6,7 @@ import com.example.e_commerce.components.ProductComponentUIModel
 import com.example.e_commerce.domain.usecase.AddProductUseCase
 import com.example.e_commerce.domain.usecase.GetProductUseCase
 import com.example.e_commerce.domain.usecase.UpdateFavoriteUseCase
-import com.example.e_commerce.utils.extension.toFavoriteLocalModel
+import com.example.e_commerce.ui.utils.mapper.toFavoriteLocalModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableSharedFlow
