+++ app/src/main/java/com/example/e_commerce/ui/utils/Util.kt
@@ -1,7 +1,8 @@
-package com.example.e_commerce.utils.util
+package com.example.e_commerce.ui.utils
 
 import android.content.Context
 
 fun dpToPx(dp: Int, context: Context): Int {
     return (dp * context.resources.displayMetrics.density).toInt()
 }
+
